+++ Contents/mods/pz-zombie-factions/42/media/lua/client/ZombieFactions/ImpactProbe_Client.lua
@@ -45,7 +45,7 @@ local processedOwnerHits = {}
 local processedOwnerHitOrder = {}
 local activeHitReactions = {}
 
-print("[ZombieFactions] Client impact probe loaded v0.0.55")
+print("[ZombieFactions] Client impact probe loaded v0.0.56")
 
 local function print(message)
     CombatController.detail(message)